Move license check logic to pkg.
