+++ lib/cli/commands/release.js
@@ -11,7 +11,6 @@ const analyse = require('./analyse'),
       files = require('../../files'),
       git = require('../../automation/git'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
-      license = require('./license'),
       pkg = require('../../automation/pkg'),
       test = require('./test');
 
@@ -95,7 +94,18 @@ const release = {
           }
         }
 
-        await license.run({});
+        try {
+          await pkg.checkLicenseCompatibility({ directory });
+        } catch (ex) {
+          switch (ex.code) {
+            case 'ELICENSEINCOMPATIBLE':
+              throw ex;
+            case 'ELICENSENOTSUPPORTED':
+              break;
+            default:
+              throw ex;
+          }
+        }
       }
 
       buntstift.line();
