+++ test/integration/deps/warns-on-outdated-dependency/expected.js
@@ -4,7 +4,7 @@ const exitCode = 0;
 
 const stdout = `▸ Checking for outdated dependencies...
 Package  Current  Wanted       Latest  Location
-noop3`;
+noop3     13.7.2  13.7.2`;
 
 const stderr = '';
