--- test/integration/release/does-not-warn-on-up-to-date-dependencies/expected.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const exitCode = 0;
-
-const stdout = `Checking for outdated dependencies...
-─`;
-
-const stderr = '';
-
-module.exports = { exitCode, stdout, stderr };
