+++ test/integration/release/fails-on-incompatible-license/pre.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const shell = require('shelljs');
-
 const helpers = require('../../../helpers');
 
 const pre = async function (options) {
@@ -10,8 +8,6 @@ const pre = async function (options) {
   await helpers.createGitRepository({
     dirname
   });
-
-  shell.exec('npm install --no-package-lock', { cwd: dirname, silent: false });
 };
 
 module.exports = pre;
