+++ test/integration/release/warns-on-outdated-dependencies/expected.js
@@ -2,8 +2,11 @@
 
 const exitCode = 0;
 
-const stdout = `Package  Current  Wanted       Latest  Location
-noop3    MISSING  13.7.2`;
+const stdout = [
+  `Checking for outdated dependencies`,
+  `Package  Current  Wanted       Latest  Location
+noop3     13.7.2  13.7.2`
+];
 
 const stderr = '';
