+++ test/integration/release/warns-on-outdated-dependencies/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const noop = require('noop3');
+
+noop();
