+++ test/integration/release/warns-on-unused-dependencies/expected.js
@@ -2,9 +2,12 @@
 
 const exitCode = 0;
 
-const stdout = `Following dependencies are unused or missing:
+const stdout = [
+  'Checking for unused or missing dependencies',
+  `Following dependencies are unused or missing:
 
-noop3`;
+noop3`
+];
 
 const stderr = ``;
